COOKgptAPI/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
drf-yasg==1.21.8
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3
Pillow==11.1.0
django-redis==5.4.0
celery==5.4.0
redis==5.2.1
mssql-django==1.5
pyodbc==5.2.0
